+++ pages/paleocene/issues/02.tsx
@@ -35,7 +35,7 @@ const Page: NextPage = () => {
             <Head favIconType="paleocene" socialImagePath="/paleocene/02" subject={SUBJECT}>
                 <ProductSchema product={PRODUCTS_MAP[PRODUCT_ID]} />
             </Head>
-            <Layout headerPromos={["paleocene"]} footerPromos={["pleistocene", "phylopic"]} theme="night">
+            <Layout headerPromos={["paleocene"]} footerPromos={["pleistocene", "parry-and-carney"]} theme="night">
                 <IssueHero
                     ageRecommendation="9+"
                     id="paleocene"