+++ pages/paleocene/issues/04.tsx
@@ -31,7 +31,7 @@ const Page: NextPage = () => {
             <Head favIconType="paleocene" socialImagePath="/paleocene/04" subject={SUBJECT}>
                 <ProductSchema product={PRODUCTS_MAP[PRODUCT_ID]} />
             </Head>
-            <Layout headerPromos={["paleocene"]} footerPromos={["pleistocene", "phylopic"]} theme="night">
+            <Layout headerPromos={["paleocene"]} footerPromos={["pleistocene", "parry-and-carney"]} theme="night">
                 <IssueHero ageRecommendation="9+" id="paleocene" number={4} seriesTitle="Paleocene">
                     <div style={{ fontSize: "1.6rem" }}>
                         <ComicText>Coming soon.</ComicText>